Model was changed

+++ src/main/java/lab3/gradebook/nc/model/Grade.java
@@ -11,6 +11,6 @@ public class Grade {
     private Person person;
     private Subject subject;
     private Lesson lesson;
-    private float grade;
+    private double grade;
 
 }
